Move legacy TimeAndSalary
