+++ src/components/Company/TimeAndSalaryScreen.js
@@ -4,33 +4,30 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import R from 'ramda';
 import qs from 'qs';
 import { compose, setStatic } from 'recompose';
-
 import Loading from 'common/Loader';
 import { P } from 'common/base';
 import FanPageBlock from 'common/FanPageBlock';
-import WorkingHourBlock from './WorkingHourBlock';
 import { withPermission } from 'common/permission-context';
-import { queryJobTitle } from '../../../actions/timeAndSalaryJobTitle';
-import { isFetching, isFetched } from '../../../constants/status';
-import renderHelmet from './helmet';
-
+import Pagination from 'common/Pagination';
 import {
   querySelector,
   pathnameSelector,
   paramsSelector,
 } from 'common/routing/selectors';
-import { pageSelector } from '../common/selectors';
-import { validatePage } from '../common/validators';
+import { queryCompany } from '../../actions/timeAndSalaryCompany';
+import { isFetching, isFetched } from '../../constants/status';
+import { pageSelector } from '../TimeAndSalary/common/selectors';
+import { validatePage } from '../TimeAndSalary/common/validators';
+import WorkingHourBlock from './WorkingHourBlock';
+import renderHelmet from './helmet';
+import styles from './TimeAndSalaryScreen.module.css';
 
-import styles from '../views/view.module.css';
-import Pagination from '../../common/Pagination/Pagination';
-
-const jobTitleSelector = R.compose(
-  params => params.jobTitle,
+const companyNameSelector = R.compose(
+  params => params.companyName,
   paramsSelector,
 );
 
-class TimeAndSalaryJobTitle extends Component {
+class TimeAndSalary extends Component {
   static propTypes = {
     data: ImmutablePropTypes.map,
     status: PropTypes.string,
@@ -39,7 +36,7 @@ class TimeAndSalaryJobTitle extends Component {
       path: PropTypes.string.isRequired,
       params: PropTypes.object.isRequired,
     }),
-    queryJobTitle: PropTypes.func,
+    queryCompany: PropTypes.func,
     history: PropTypes.shape({
       push: PropTypes.func.isRequired,
     }).isRequired,
@@ -48,16 +45,16 @@ class TimeAndSalaryJobTitle extends Component {
   };
 
   componentDidMount() {
-    this.props.queryJobTitle({
-      jobTitle: jobTitleSelector(this.props),
+    this.props.queryCompany({
+      companyName: companyNameSelector(this.props),
     });
     this.props.fetchPermission();
   }
 
   componentDidUpdate(prevProps) {
-    if (jobTitleSelector(prevProps) !== jobTitleSelector(this.props)) {
-      this.props.queryJobTitle({
-        jobTitle: jobTitleSelector(this.props),
+    if (companyNameSelector(prevProps) !== companyNameSelector(this.props)) {
+      this.props.queryCompany({
+        companyName: companyNameSelector(this.props),
       });
       this.props.fetchPermission();
     }
@@ -69,14 +66,14 @@ class TimeAndSalaryJobTitle extends Component {
     const page = validatePage(pageSelector(this.props));
     const pageSize = 10;
 
-    const jobTitle = jobTitleSelector(this.props);
-    const title = `${jobTitle} 薪水、加班情況`;
+    const companyName = companyNameSelector(this.props);
+    const title = `${companyName} 薪水、加班情況`;
 
     const queryParams = querySelector(this.props);
 
     return (
       <section className={styles.searchResult}>
-        {renderHelmet({ title, pathname, jobTitle })}
+        {renderHelmet({ title, pathname, companyName })}
         <h2 className={styles.heading}>{title}</h2>
         {isFetching(status) && <Loading size="s" />}
         {isFetched(status) &&
@@ -106,8 +103,8 @@ class TimeAndSalaryJobTitle extends Component {
             </React.Fragment>
           )) || (
             <P size="l" bold className={styles.searchNoResult}>
-              尚未有職稱「
-              {jobTitle}
+              尚未有公司「
+              {companyName}
               」的薪時資訊
             </P>
           ))}
@@ -117,9 +114,9 @@ class TimeAndSalaryJobTitle extends Component {
 }
 
 const ssr = setStatic('fetchData', ({ store: { dispatch }, ...props }) => {
-  const jobTitle = jobTitleSelector(props);
+  const companyName = companyNameSelector(props);
 
-  return dispatch(queryJobTitle({ jobTitle }));
+  return dispatch(queryCompany({ companyName }));
 });
 
 const hoc = compose(
@@ -127,4 +124,4 @@ const hoc = compose(
   withPermission,
 );
 
-export default hoc(TimeAndSalaryJobTitle);
+export default hoc(TimeAndSalary);
