+++ src/components/Company/TimeAndSalaryScreen.module.css
@@ -67,38 +67,6 @@ $bold: 700;
   margin-top: 40px;
 }
 
-.sort {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-bottom: 20px;
-
-  @media (max-width: $below-mobile) {
-    text-align: left;
-    display: flex;
-    align-items: center;
-  }
-  .label {
-    display: inline-block;
-    position: relative;
-    top: 3px;
-
-    @media (max-width: $below-mobile) {
-      flex: 0 0 65px;
-    }
-  }
-  .select {
-    display: inline-block;
-    max-width: 225px;
-    border-color: $gray;
-
-    @media (max-width: $below-mobile) {
-      max-width: initial;
-      flex: 1 1 auto;
-    }
-  }
-}
-
 .heading {
   @mixin heading-style;
   text-align: center;
@@ -106,39 +74,6 @@ $bold: 700;
   @media (max-width: $below-mobile) {
     margin-bottom: $block-gutter-s;
   }
-}
-
-.bar {
-  background-color: yellow-bar;
-  padding: 5px;
-}
-
-.dot {
-  display: inline-block;
-  border-radius: 50%;
-  width: 10px;
-  height: 10px;
-  margin-right: 10px;
-  &.hardly {
-    background-color: main-gray;
-  }
-  &.sometimes {
-    background-color: gray-dark;
-  }
-  &.usually {
-    background-color: main-yellow;
-  }
-  &.always {
-    background-color: warning-red;
-  }
-}
-
-.sector {
-  color: main-gray;
-}
-
-.feedbackBlock {
-  margin-top: 60px;
 }
 
 .fanPageBlock {
