+++ src/components/Company/WorkingHourBlock.js
@@ -2,9 +2,10 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 import { Heading } from 'common/base';
+import OvertimeBlock from '../TimeAndSalary/common/OvertimeBlock';
 import WorkingHourTable from './WorkingHourTable';
 
-import styles from '../common/WorkingHourBlock.module.css';
+import styles from '../TimeAndSalary/common/WorkingHourBlock.module.css';
 
 class WorkingHourBlock extends Component {
   static propTypes = {
@@ -16,6 +17,13 @@ class WorkingHourBlock extends Component {
     const { data, hideContent } = this.props;
     return (
       <div>
+        <div className={styles.overtimeBlock}>
+          <div className={styles.overtimeBlockInner}>
+            <OvertimeBlock type="salary" heading="加班有無加班費" data={data} />
+            <OvertimeBlock type="dayoff" heading="加班有無補休" data={data} />
+          </div>
+          <div className={styles.overtimeBlockUnit}>單位：資料筆數</div>
+        </div>
         <WorkingHourTable
           data={data.salary_work_times}
           hideContent={hideContent}
