+++ src/components/TimeAndSalary/SearchScreen/index.js
@@ -1,0 +1,3 @@
+import SearchScreen from './SearchScreen';
+
+export default SearchScreen;
