+++ src/components/TimeAndSalary/common/formatter.module.css
@@ -1,0 +1,30 @@
+@value yellow-bar, gray-dark, main-gray, main-yellow, warning-red from "common/variables.module.css";
+
+.bar {
+  background-color: yellow-bar;
+  padding: 5px;
+}
+
+.dot {
+  display: inline-block;
+  border-radius: 50%;
+  width: 10px;
+  height: 10px;
+  margin-right: 10px;
+  &.hardly {
+    background-color: main-gray;
+  }
+  &.sometimes {
+    background-color: gray-dark;
+  }
+  &.usually {
+    background-color: main-yellow;
+  }
+  &.always {
+    background-color: warning-red;
+  }
+}
+
+.sector {
+  color: main-gray;
+}
